+++ src/floorplanner/floorplanner_view.ts
@@ -1,20 +1,14 @@
 /// <reference path="../../lib/jQuery.d.ts" />
+/// <reference path="../core/configuration.ts" />
 /// <reference path="../model/floorplan.ts" />
 /// <reference path="../model/half_edge.ts" />
 /// <reference path="../model/model.ts" />
 /// <reference path="../model/wall.ts" />
+/// <reference path="../utils/dimensioning.ts" />
 /// <reference path="../utils/utils.ts" />
 /// <reference path="floorplanner.ts" />
 
 module BP3D.Floorplanner {
-  /** Converts cm to feet string. */
-  const cmToFeet: ((number) => string) = (cm) => {
-    var realFeet = ((cm * 0.393700) / 12);
-    var feet = Math.floor(realFeet);
-    var inches = Math.round((realFeet - feet) * 12);
-    return feet + "'" + inches + '"';
-  }
-
   /** */
   export const floorplannerModes = {
     MOVE: 0,
@@ -161,10 +155,10 @@ module BP3D.Floorplanner {
       this.context.strokeStyle = "#ffffff";
       this.context.lineWidth = 4;
 
-      this.context.strokeText(cmToFeet(length),
+      this.context.strokeText(Dimensioning.cmToMeasure(length),
         this.viewmodel.convertX(pos.x),
         this.viewmodel.convertY(pos.y));
-      this.context.fillText(cmToFeet(length),
+      this.context.fillText(Dimensioning.cmToMeasure(length),
         this.viewmodel.convertX(pos.x),
         this.viewmodel.convertY(pos.y));
     }
